[feature] buildSrc로 함수 따로 관리
