+++ buildSrc/build.gradle.kts
@@ -1,0 +1,9 @@
+import org.gradle.kotlin.dsl.`kotlin-dsl`
+
+plugins {
+    `kotlin-dsl`
+}
+repositories {
+    google()
+    mavenCentral()
+}
